+++ wasmc/src/wasm/Directives.cpp
@@ -58,7 +58,7 @@ namespace Wasmc {
 		delete value_size;
 	}
 
-	AlignDirective::AlignDirective(long alignment_): Directive(WASM_ALIGNDIR), alignment(alignment_) {}
+	AlignDirective::AlignDirective(size_t alignment_): Directive(WASM_ALIGNDIR), alignment(alignment_) {}
 
 	FillDirective::FillDirective(long count_, long value_): Directive(WASM_FILLDIR), count(count_), value(value_) {}
